LRCService/requirements.txt
tensorflow
numpy
tensorflow_datasets
grpcio